10.18 오류처리 및 페이지 흐름 제어 추가

+++ routes/upload/upload.js
@@ -6,30 +6,36 @@ var s3 = new aws.S3();
 var connection = require('../connection');
 
 router.get('/', (req, res) => {
-    res.render('upload');
+    if (req.session.user_info == undefined) res.redirect('/error/info');
+    else {
+        res.render('upload');
+    }
 });
 
 router.post('/', upload.single('image'), (req, res) => {
     sql = 'update user set image = 1 where user_id = ?'
 
-    connection.query(sql, req.user, function(err, result) {
-        if (err) throw err;
+    connection.query(sql, req.user, function (err, result) {
+        if (err) {console.error(err); res.redirect('/error/connect')}
         res.redirect('/profile');
     });
 });
 
 router.get('/del', (req, res) => {
-    s3.deleteObject({
-        Bucket: process.env.BUCKET_NAME,
-        Key: 'profile/' + req.user
-    })
+    if (req.session.user_info == undefined) res.redirect('/error/info');
+    else {
+        s3.deleteObject({
+            Bucket: process.env.BUCKET_NAME,
+            Key: 'profile/' + req.user
+        })
 
-    sql = 'update user set image = null where user_id = ?'
+        sql = 'update user set image = null where user_id = ?'
 
-    connection.query(sql, req.user, function(err, result) {
-        if (err) throw err;
-        res.redirect('/profile');
-    });
+        connection.query(sql, req.user, function (err, result) {
+            if (err) {console.error(err); res.redirect('/error/connect')}
+            res.redirect('/profile');
+        });
+    }
 });
 
 module.exports = router;
